Coverage entre 53% e 64%
