+++ semana-11/aula-3/planner/package.json
@@ -5,11 +5,13 @@
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
+    "@testing-library/react-hooks": "^3.3.0",
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.19.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.0",
+    "react-test-renderer": "^16.13.1",
     "styled-components": "^5.1.1"
   },
   "scripts": {
