+++ semana-11/aula-3/planner/src/components/Header/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Form from './Form';
+import { SHeader } from './styles';
+
+const Header = ({ fields, setFields, handleSubmit }) => {
+  return (
+    <SHeader>
+      <div>
+        <h1>Weekly Planner</h1>
+      </div>
+      <div>
+        <Form {...{ fields, setFields, handleSubmit }} />
+      </div>
+    </SHeader>
+  );
+};
+
+export default Header;
