+++ semana-11/aula-3/planner/src/components/Header/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const SHeader = styled.header`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 32px 0;
+  > div {
+    padding: 0 16px;
+  }
+`;
+
+export { SHeader };
